Move FaceManagementService together with the others services
