+++ lib/Service/FaceManagementService.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @copyright Copyright (c) 2017, Matias De lellis <mati86dl@gmail.com>
+ * @copyright Copyright (c) 2017, 2020 Matias De lellis <mati86dl@gmail.com>
  * @copyright Copyright (c) 2018, Branko Kokanovic <branko@kokanovic.org>
  *
  * @author Branko Kokanovic <branko@kokanovic.org>
@@ -21,9 +21,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-namespace OCA\FaceRecognition;
+namespace OCA\FaceRecognition\Service;
 
-use OCP\IConfig;
 use OCP\IUser;
 use OCP\IUserManager;
 
@@ -31,7 +30,7 @@ use OCA\FaceRecognition\Db\FaceMapper;
 use OCA\FaceRecognition\Db\ImageMapper;
 use OCA\FaceRecognition\Db\PersonMapper;
 
-use OCA\FaceRecognition\BackgroundJob\Tasks\AddMissingImagesTask;
+use OCA\FaceRecognition\Service\SettingService;
 
 /**
  * Background service. Both command and cron job are calling this service for long-running background operations.
@@ -43,9 +42,6 @@ use OCA\FaceRecognition\BackgroundJob\Tasks\AddMissingImagesTask;
  * reason about them and test them independently. Other than that, they are really glorified functions.
  */
 class FaceManagementService {
-
-	/** @var IConfig */
-	private $config;
 
 	/** @var IUserManager */
 	private $userManager;
@@ -59,13 +55,20 @@ class FaceManagementService {
 	/** @var PersonMapper */
 	private $personMapper;
 
-	public function __construct(IConfig $config, IUserManager $userManager,
-			FaceMapper $faceMapper, ImageMapper $imageMapper, PersonMapper $personMapper) {
-		$this->config = $config;
-		$this->userManager = $userManager;
-		$this->faceMapper = $faceMapper;
-		$this->imageMapper = $imageMapper;
-		$this->personMapper = $personMapper;
+	/** @var SettingService */
+	private $settingService;
+
+	public function __construct(IUserManager   $userManager,
+	                            FaceMapper     $faceMapper,
+	                            ImageMapper    $imageMapper,
+	                            PersonMapper   $personMapper,
+	                            SettingService $settingService)
+	{
+		$this->userManager    = $userManager;
+		$this->faceMapper     = $faceMapper;
+		$this->imageMapper    = $imageMapper;
+		$this->personMapper   = $personMapper;
+		$this->settingService = $settingService;
 	}
 
 	/**
@@ -97,6 +100,7 @@ class FaceManagementService {
 		$this->faceMapper->deleteUserFaces($userId);
 		$this->personMapper->deleteUserPersons($userId);
 		$this->imageMapper->deleteUserImages($userId);
-		$this->config->deleteUserValue($userId, 'facerecognition', AddMissingImagesTask::FULL_IMAGE_SCAN_DONE_KEY);
+		$this->settingService->setUserFullScanDone(false, $userId);
 	}
+
 }
